requirements.txt
Django==3.0.5
psycopg2==2.8.5
django-environ
djangorestframework
markdown
django-filter
